+++ wagtail_api_forms/urls.py
@@ -5,14 +5,13 @@ from django.conf.urls.i18n import i18n_patterns
 from django.urls import include, path
 from django.contrib import admin
 
-from wagtail.admin import urls as wagtailadmin_urls
 from wagtail import urls as wagtail_urls
+from wagtail.admin import urls as wagtailadmin_urls
 from wagtail.documents import urls as wagtaildocs_urls
 
-from sphinx_view import DocumentationView
-
 from wagtail_api_forms.formpages import urls as formpages_urls
-from wagtail_api_forms.home.views import sphinx_searchindex_view
+# from sphinx_view import DocumentationView
+# from wagtail_api_forms.home.views import sphinx_searchindex_view
 
 
 urlpatterns = [
@@ -26,15 +25,15 @@ urlpatterns = [
     path('cspreports/', include('cspreports.urls')),
 
     # django_sphinx_view
-    path('docs/searchindex.js', sphinx_searchindex_view, name='sphinx_searchindex_view'),
-    path(
-        "docs<path:path>",
-        DocumentationView.as_view(
-            json_build_dir=Path('docs/_build/json/'),
-            base_template_name="sphinx_view/base.html",
-        ),
-        name="documentation",
-    ),
+    # path('docs/searchindex.js', sphinx_searchindex_view, name='sphinx_searchindex_view'),
+    # path(
+    #     "docs<path:path>",
+    #     DocumentationView.as_view(
+    #         json_build_dir=Path('docs/_build/json/'),
+    #         base_template_name="sphinx_view/base.html",
+    #     ),
+    #     name="documentation",
+    # ),
 ]
 
 # Translatable URLs